third_party/blink/web_tests/external/wpt/html/syntax/parsing/html5lib_webkit02_run_type=write_single-expected.txt

This is a testharness.js-based test.
[FAIL] html5lib_webkit02.html 6e33515b4dc011dd390d433a6358bf68b786b1fd
  assert_equals: expected "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <b>\\n|       <em>\\n|         <foo>\\n|           <foo>\\n|             <foo>\\n|     <aside>\\n|       <b>" but got "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <b>\\n|       <em>\\n|         <foo>\\n|           <foo>\\n|             <foo>\\n|     <aside>\\n|       <em>\\n|         <b>"
[FAIL] html5lib_webkit02.html 4879f476053094cf5602d325724675378856a902
  assert_equals: expected "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <select>\\n|       <option>\\n|       <hr>" but got "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <select>\\n|       <option>\\n|         <hr>"
[FAIL] html5lib_webkit02.html 44c88b90236f01ebc8e0123363b527640a07070c
  assert_equals: expected "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <select>\\n|       <optgroup>\\n|         <option>\\n|       <hr>" but got "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <select>\\n|       <optgroup>\\n|         <option>\\n|           <hr>"
[FAIL] html5lib_webkit02.html cfb304e8f2d3cbdecc362226e7775cab452d5489
  assert_equals: expected "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <select>\\n|       <optgroup>\\n|       <hr>" but got "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <select>\\n|       <optgroup>\\n|         <hr>"
[FAIL] html5lib_webkit02.html 3fc625e7cb9b6ea72a9e252ede84c6fdd9680d87
  assert_equals: expected "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <select>\\n|       <option>\\n|       <optgroup>\\n|       <hr>" but got "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <select>\\n|       <option>\\n|       <optgroup>\\n|         <hr>"
[FAIL] html5lib_webkit02.html cee2230c74671c594a1140a68d16e3d3e5ae005a
  assert_equals: expected "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <table>\\n|       <tbody>\\n|         <tr>\\n|           <td>\\n|             <select>\\n|               <option>\\n|               <hr>" but got "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <table>\\n|       <tbody>\\n|         <tr>\\n|           <td>\\n|             <select>\\n|               <option>\\n|                 <hr>"
[FAIL] html5lib_webkit02.html 22b9fe36797d70a3b71a6aadc6ad7cff23c3fc90
  assert_equals: expected "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <table>\\n|       <tbody>\\n|         <tr>\\n|           <td>\\n|             <select>\\n|               <optgroup>\\n|                 <option>\\n|               <hr>" but got "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <table>\\n|       <tbody>\\n|         <tr>\\n|           <td>\\n|             <select>\\n|               <optgroup>\\n|                 <option>\\n|                   <hr>"
[FAIL] html5lib_webkit02.html a82c3bf49c381b5f58c5c8a4bbbe0cef2458e28a
  assert_equals: expected "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <table>\\n|       <tbody>\\n|         <tr>\\n|           <td>\\n|             <select>\\n|               <optgroup>\\n|               <hr>" but got "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <table>\\n|       <tbody>\\n|         <tr>\\n|           <td>\\n|             <select>\\n|               <optgroup>\\n|                 <hr>"
[FAIL] html5lib_webkit02.html 61f8d527795dc8044a95a3e2437de81e16597ceb
  assert_equals: expected "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <table>\\n|       <tbody>\\n|         <tr>\\n|           <td>\\n|             <select>\\n|               <option>\\n|               <optgroup>\\n|               <hr>" but got "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <table>\\n|       <tbody>\\n|         <tr>\\n|           <td>\\n|             <select>\\n|               <option>\\n|               <optgroup>\\n|                 <hr>"
Harness: the test ran to completion.
